Add an avatar and create an about component
