+++ src/css/Contact.scss
@@ -1,6 +1,4 @@
 .contact {
-  margin-top: 1rem;
-
   .icon {
     font-size: 2rem;
     margin: 0 0.25rem 0 0.25rem;
