+++ src/css/Header.scss
@@ -1,3 +1,4 @@
 h1 {
-  margin: 2rem 0 2.5rem 0;
+  margin-top: 2rem;
+  margin-bottom: 1.5rem;
 }
