+++ src/css/ProjectList.scss
@@ -1,7 +1,3 @@
 .row {
   justify-content: center;
 }
-
-.projectList {
-  padding: 2rem 0 2rem 0;
-}
